Add multimodal model mixins for the helpers every VLM repeats
Multimodal models re-declare the same handful of methods almost verbatim.
Counting identical bodies across the 71 modeling files that define
`get_placeholder_mask`: `get_vision_position_ids` was 17 copies of one
function, `compute_3d_position_ids` 14 of 17, `_prepare_position_ids_for_
generation` 14 of 16, `_get_image_nums_and_video_nums` 10 of 17.

Add `modeling_multimodal_utils.py` with two mixins, one per attachment
point, since these helpers live on two different classes:

- `MultiModalPreTrainedModelMixin` on the base `<X>Model`:
  `get_vision_position_ids` (now a thin method form of
  `modeling_rope_utils.get_mrope_vision_positions`, which is the same
  function it had been copied from), `get_rope_index`, and
  `compute_3d_position_ids`.
- `MultiModalGenerationMixin` on `<X>ForConditionalGeneration`, mixed in
  before `GenerationMixin` so its overrides win while `super()` still
  reaches the text-only behaviour: `_prepare_position_ids_for_generation`
  and `_get_image_nums_and_video_nums`.

Mixing them into the family roots and regenerating cascades to every
child, so this deletes the copies rather than relocating them. Families
that genuinely differ keep their overrides: qwen2_5_vl (threads
`second_per_grid_ts`), hunyuan_vl (image-only layout), the two Omni
models (audio signatures), glm_image (caches decode positions on the
module), exaone4_5 (BC shim), video_llama_3, and the glm4v family for
`_get_image_nums_and_video_nums` (different token ids).

`get_rope_index` on the mixin swallows `**kwargs` without forwarding
them, matching the per-model wrappers it replaces — generation passes the
whole model-kwargs dict through it.

Two things a mixin cannot do that a parent class body can, both of which
bit during this refactor and are handled at the mixin level rather than
per model:

- It cannot be *deleted* by modular's `raise AttributeError` idiom. Four
  models used that to drop these helpers (video_llama_3, exaone4_5,
  kimi_k25, muse_glimmer) and the stubs started being emitted into
  generated code, where `generate` would have called them. Those four are
  exactly the models whose configs declare no M-RoPE, so the mixin now
  gates on `uses_mrope(self.config)`: `compute_3d_position_ids` returns
  `None` and `_prepare_position_ids_for_generation` returns the text
  positions, which is what an absent method used to give them. The stubs
  are gone.
- It changes what `super()` means in existing overrides. cosmos3_edge and
  hunyuan_vl start from the plain 2D text positions, and their `super()`
  call silently began returning the mixin's four-axis result instead
  (caught by five cosmos3_edge generate tests). Rather than pin those two
  call sites, `_prepare_position_ids_for_generation` is now a thin hook
  that hands `text_positions` to an overridable
  `_prepare_mrope_position_ids_for_generation`, so no override has to
  reason about the MRO.
