+++ src/transformers/models/kimi_k25/modular_kimi_k25.py
@@ -708,9 +708,6 @@ class Kimi_K25ForConditionalGeneration(Glm4vForConditionalGeneration):
             attentions=outputs.attentions,
         )
 
-    def _prepare_position_ids_for_generation(self, **kwargs):
-        raise AttributeError("Kimi doesn't use m-rope!")
-
     def _get_image_nums_and_video_nums(self, **super_kwargs):
         raise AttributeError()
